requirements.txt
asgiref==3.8.1
asttokens==2.4.1
boto3==1.35.7
botocore==1.35.7
decorator==5.1.1
Django==5.1
django-storages==1.14.4
djangorestframework==3.15.2
executing==2.0.1
ipython==8.26.0
jedi==0.19.1
jmespath==1.0.1
matplotlib-inline==0.1.7
parso==0.8.4
pexpect==4.9.0
prompt_toolkit==3.0.47
psycopg2==2.9.9
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.18.0
python-dateutil==2.9.0.post0
s3transfer==0.10.2
six==1.16.0
sqlparse==0.5.1
stack-data==0.6.3
traitlets==5.14.3
urllib3==2.2.2
wcwidth==0.2.13
